Increase max bulk_batch_size
The previous limit was a max of 1000 query parameters. This is changed
to a max of 1000 rows (the max number of rows allowed for inserting in a
single statement by mssql) or however many rows can be inserted while
keeping the number of query parameters below 2050 per insert statement
(mssql reports a max allowed of 2100 parameters but a few parameters are
reserved for executing the query).