Don't drop unique indexes before M2M rename
This method override was added by:
https://github.com/ESSolutions/django-mssql-backend/pull/24/commits/08c2721a681062eccbb0c2cd33a9cf5661ec92f7
meaning that it dropped (but did not re-instate) any unique
index on the M2M through table, before renaming that table.

However MSSQL does not seem to require this, so rather
than trying to re-create the correct indexes afterwards
just don't drop them unnecessarily in the first place.