Fix Now() to be timezone-aware
When USE_TZ=True, it's important that Now() reflect the current time, with timezone offset. If the database system OS is not using UTC , the SYSDATETIME() is a local-timezone value, which incorrectly CASTED as UTC (because offset = 00:00). This fix and the previous one allows to leverage the timezone part of SYSDATETIMEOFFSET() to ensure the timezone offset makes Now() the correct value.